# public/openapi.yaml
openapi: 3.1.0
info:
  title: Juniper Commerce — onX API
  description: |
    REST API for the Juniper Commerce reference implementation of the onX (Order Network eXchange) standard.
    Supports all 12 standard onX operations covering products, inventory, orders, fulfillments, and returns.
  version: 2.0.0
  contact:
    name: Commerce Operations Foundation
    url: https://commerceopsfoundation.org
  license:
    name: MIT
servers:
  - url: /api/onx
    description: Local onX API
paths:
  /products:
    get:
      summary: Get Products
      operationId: getProducts
      tags: [Products]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive] }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: includeVariants
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Product' } }
                  total: { type: integer }
  /product-variants:
    get:
      summary: Get Product Variants
      operationId: getProductVariants
      tags: [Products]
      parameters:
        - name: productId
          in: query
          schema: { type: string }
        - name: sku
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Variant list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ProductVariant' } }
                  total: { type: integer }
...
  /orders:
    get:
      summary: Get Orders
      operationId: getOrders
      tags: [Orders]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [confirmed, processing, shipped, delivered, cancelled] }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Order list
    post:
      summary: Create Sales Order
      operationId: createSalesOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderInput' }
      responses:
        '201':
          description: Order created
...
components:
  schemas:
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        status: { type: string }
        tags: { type: array, items: { type: string } }
        categories: { type: array, items: { type: string } }
        options: { type: array, items: { type: object } }
        imageURLs: { type: array, items: { type: string } }
    ProductVariant:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        sku: { type: string }
        title: { type: string }
        price: { type: number }
        currency: { type: string }
...
    CreateOrderInput:
      type: object
      required: [customer, lineItems, shippingAddress]
      properties:
        customer:
          type: object
          required: [email, firstName, lastName]
          properties:
            email: { type: string }
            firstName: { type: string }
            lastName: { type: string }
        lineItems:
          type: array
          items:
            type: object
            required: [sku, quantity]
            properties:
              sku: { type: string }
              quantity: { type: integer }
        shippingAddress:
          type: object
          required: [address1, city, stateOrProvince, zipCodeOrPostalCode, country]
          properties:
            address1: { type: string }
            city: { type: string }
            stateOrProvince: { type: string }
            zipCodeOrPostalCode: { type: string }
            country: { type: string }
...